#28, #29 - create a layout for all modal
